requirements.txt
absl-py==0.7.1
argh==0.26.2
astor==0.8.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
joblib==0.13.2
jsonrpclib-pelix==0.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
nltk==3.4.4
numpy==1.16.4
pandas==0.25.0
pathtools==0.1.2
pika==1.0.1
pkg-resources==0.0.0
protobuf==3.9.0
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
watchdog==0.9.0
Werkzeug==0.15.5
wrapt==1.11.2
